requirements.txt
astroid==2.3.2
autopep8==1.4.4
certifi==2019.9.11
chardet==3.0.4
colorama==0.4.1
Django==2.2.7
django-rest-framework==0.1.0
django-smtp-ssl==1.0
django-user-agents==0.4.0
django-webpack-loader==0.6.0
djangorestframework==3.10.3
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
mysqlclient==1.4.5
pycodestyle==2.5.0
pylint==2.4.3
pylint-django==2.0.11
pylint-plugin-utils==0.6
python-dotenv==0.10.3
pytz==2019.3
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
ua-parser==0.8.0
urllib3==1.25.7
user-agents==2.0
whitenoise==4.1.4
wrapt==1.11.2
xkcdpass==1.17.3
